pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pydjirecord"
version = "0.1.0"
description = "Python parser for DJI drone flight log files"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "httpx>=0.27",
    "pycryptodome>=3.20",
]

[project.scripts]
djirecord = "pydjirecord.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["pydjirecord"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["Crypto.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=pydjirecord --cov-report=term-missing -v"

[tool.coverage.run]
source = ["pydjirecord"]

[tool.coverage.report]
fail_under = 80

[tool.coverage.paths]
source = ["src/pydjirecord"]
